LPD-15510 Shared LicenseTerms Component
